docs: added more details to example in readme.md
