+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './budgetNotifier';
+export * from './budgetNotifierProps';
 export * from './notificationType';
 export * from './timeUnit';
